+++ pruebas/01-reading-list/jordislpm/src/Hooks/useSearchBook.jsx
@@ -38,7 +38,7 @@ const useSearchBook = ()=>{
 
 
 
-    return[search, setSearch , error]
+    return{search, setSearch , error}
 
 }
