+++ pruebas/01-reading-list/jordislpm/src/components/BookList/bookList.module.css
@@ -19,3 +19,7 @@
     font-size: 2rem;
     color: var(--secundary-blue)
 }
+
+.notFound{
+    color: var(--primary-red);
+}
