+++ pruebas/01-reading-list/jordislpm/src/components/Header/header.module.css
@@ -14,7 +14,8 @@
 .availables{
     font-weight: bold;
     font-size: 2.5rem;
-    color: var(--secundary-blue)
+    color: var(--secundary-blue);
+    text-align: center;
 }
 
 .read{
